Optimize deal cover photos with next/image
Configure next.config.js to allow next/image to serve optimized webp
images (instead of images.unoptimized passthrough), whitelisting the S3
upload bucket and the IPFS gateways already used elsewhere in the app
(see IPFS_GATEWAYS in backend/.env.example) as remote patterns.

Extract the inline DealCard from the marketplace grid into its own
frontend/src/components/marketplace/DealCard.tsx, rendering the cover
photo through next/image with a sizes attribute matching the card
grid's breakpoints so the browser fetches an appropriately sized image
per viewport rather than a full-resolution upload. No deal currently
has a cover photo in the schema, so the Deal type gains an optional
cover_image_url field and the card falls back to a placeholder when
it's absent.

Closes #266
